requirements.txt
fastapi==0.135.1
uvicorn[standard]>=0.30.0
python-jobspy==1.1.82
pydantic==2.12.5
pyyaml>=6.0.2
